requirements.txt
absl-py==1.2.0
astunparse==1.6.3
bson==0.5.10
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.19.5
importlib-metadata==4.12.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.4.1
MarkupSafe==2.1.1
networkx==2.8.5
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tifffile==2022.5.4
typing_extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
